+++ components/common/dropdowns/dropdown.tsx
@@ -16,7 +16,7 @@ import {
   pickCommentOptions,
   pickpickpickDropdownOptions,
   techBlogDropdownOptions,
-} from '@/constants/DropdownOptionArr';
+} from '@/constants/DropdownOption';
 import {
   DropdownOptionProps,
   useBlameReasonStore,
